refactor(task-4.2): extract image constraints
