+++ web/src/services/api/image-constraints.ts
@@ -1,0 +1,103 @@
+export const IMAGE_SIZE_OPTIONS = ["auto", "1:1", "4:3", "3:4", "16:9", "9:16"] as const;
+export const MAX_IMAGE_RATIO = 3;
+export const IMAGE_OUTPUT_FORMAT = "png";
+
+const QUALITY_BASE: Record<string, number> = {
+    low: 1024,
+    medium: 2048,
+    high: 2880,
+    standard: 1024,
+    hd: 2048,
+};
+
+const QUALITY_ALIASES: Record<string, string> = {
+    "1k": "low",
+    "2k": "medium",
+    "4k": "high",
+};
+
+const DEFAULT_IMAGE_SHORT_SIDE = 1024;
+const IMAGE_SIZE_STEP = 16;
+const IMAGE_MIN_PIXELS = 655360;
+const IMAGE_MAX_PIXELS = 8294400;
+const IMAGE_MAX_EDGE = 3840;
+
+export function normalizeImageQuality(quality: string) {
+    const value = quality.trim().toLowerCase();
+    const normalized = QUALITY_ALIASES[value] || value;
+    return QUALITY_BASE[normalized] ? normalized : undefined;
+}
+
+export function parseImageRatio(size: string) {
+    const value = size.trim();
+    if (!value || value.toLowerCase() === "auto") return null;
+    const parts = value.split(":");
+    if (parts.length !== 2) return null;
+    const [rawWidth, rawHeight] = parts.map(Number);
+    if (!Number.isFinite(rawWidth) || !Number.isFinite(rawHeight) || rawWidth <= 0 || rawHeight <= 0) return null;
+    const ratio = rawWidth / rawHeight;
+    if (ratio <= 0 || ratio > MAX_IMAGE_RATIO || 1 / ratio > MAX_IMAGE_RATIO) return null;
+    return ratio;
+}
+
+export function resolveImageRequestSize(quality: string | undefined, size: string) {
+    const value = size.trim();
+    if (!value || value.toLowerCase() === "auto") return undefined;
+    const dimensions = parseImageDimensions(value);
+    if (dimensions) {
+        validateImageSize(dimensions.width, dimensions.height);
+        return `${dimensions.width}x${dimensions.height}`;
+    }
+    if (value.includes(":")) return resolveSize(quality, value);
+    throw new Error("图像尺寸格式不支持，请使用 auto、9:16 或 1024x1024");
+}
+
+/** Map "quality + ratio" to an explicit pixel dimension like "3840x2160". */
+function resolveSize(quality: string | undefined, ratio: string): string {
+    const parsedRatio = parseImageRatioDimensions(ratio);
+    const basePixels = quality ? QUALITY_BASE[quality] : undefined;
+    const isLandscape = parsedRatio.width >= parsedRatio.height;
+    const longRatio = isLandscape ? parsedRatio.width / parsedRatio.height : parsedRatio.height / parsedRatio.width;
+    let longSide: number;
+    let shortSide: number;
+
+    if (basePixels) {
+        const targetPixels = basePixels * basePixels;
+        const longSideRaw = Math.sqrt(targetPixels * longRatio);
+        longSide = Math.floor(longSideRaw / IMAGE_SIZE_STEP) * IMAGE_SIZE_STEP;
+        shortSide = Math.round(longSide / longRatio / IMAGE_SIZE_STEP) * IMAGE_SIZE_STEP;
+    } else {
+        shortSide = DEFAULT_IMAGE_SHORT_SIDE;
+        longSide = Math.round((shortSide * longRatio) / IMAGE_SIZE_STEP) * IMAGE_SIZE_STEP;
+    }
+
+    const width = isLandscape ? longSide : shortSide;
+    const height = isLandscape ? shortSide : longSide;
+    validateImageSize(width, height);
+    return `${width}x${height}`;
+}
+
+function parseImageRatioDimensions(value: string) {
+    const parts = value.split(":");
+    if (parts.length !== 2) throw new Error("图像尺寸格式不支持，请使用 auto、9:16 或 1024x1024");
+    const w = Number(parts[0]);
+    const h = Number(parts[1]);
+    if (!Number.isFinite(w) || !Number.isFinite(h) || w <= 0 || h <= 0) throw new Error("图像比例必须是正数，例如 9:16");
+    if (Math.max(w, h) / Math.min(w, h) > MAX_IMAGE_RATIO) throw new Error("图像宽高比不能超过 3:1，请调整尺寸");
+    return { width: w, height: h };
+}
+
+function parseImageDimensions(value: string) {
+    const match = value.match(/^(\d+)x(\d+)$/i);
+    if (!match) return null;
+    return { width: Number(match[1]), height: Number(match[2]) };
+}
+
+function validateImageSize(width: number, height: number) {
+    if (!Number.isInteger(width) || !Number.isInteger(height) || width <= 0 || height <= 0) throw new Error("图像尺寸必须是正整数，例如 1024x1024");
+    if (width % IMAGE_SIZE_STEP !== 0 || height % IMAGE_SIZE_STEP !== 0) throw new Error("图像尺寸的宽高必须是 16 的倍数，请调整尺寸");
+    if (Math.max(width, height) > IMAGE_MAX_EDGE) throw new Error("图像尺寸最长边不能超过 3840px，请调整尺寸");
+    if (Math.max(width, height) / Math.min(width, height) > MAX_IMAGE_RATIO) throw new Error("图像宽高比不能超过 3:1，请调整尺寸");
+    const pixels = width * height;
+    if (pixels < IMAGE_MIN_PIXELS || pixels > IMAGE_MAX_PIXELS) throw new Error("图像总像素需在 655360 到 8294400 之间，请调整尺寸");
+}
